requirements.txt
pkg-resources==0.0.0
pycryptodome==3.9.0
urllib3==1.25.6
xmltodict==0.12.0
